+++ notebooks/preview.py
@@ -1,6 +1,5 @@
 # %%
 import ee
-import matplotlib.pyplot as plt
 import pandas as pd
 import geemap
 import ipywidgets as widgets
@@ -8,7 +7,6 @@ from IPython.display import display
 import sys
 import os
 from tqdm import tqdm
-from datetime import timedelta
 from flaring_evaluation import evaluateScene
 
 mbsp_path = os.path.join(
@@ -77,6 +75,9 @@ def add_glint_alpha(sid_data, collection="COPERNICUS/S2_HARMONIZED"):
 
 
 def calculate_sunglint_alpha(image: ee.Image):
+    """
+    Calculates the glint alpha metadata for estimating sun glint in a scene
+    """
     pi = 3.141592653589793
 
     # --- Solar geometry (radians) ---
@@ -121,106 +122,31 @@ def calculate_sunglint_alpha(image: ee.Image):
     return image.set("glint_alpha", alpha_deg)
 
 
-def evaluate_offshore_site(image, lat, lon, buffer_m=200):
-    # Define point and buffer
-    point = ee.Geometry.Point([lon, lat])
-    region = point.buffer(buffer_m)
-
-    # Select relevant bands from S2
-    b3 = image.select("B3")  # Green
-    b8 = image.select("B8")  # NIR
-    b11 = image.select("B11")  # SWIR1
-    b12 = image.select("B12")  # SWIR2
-    b8a = image.select("B8A")  # Narrow NIR
-
-    # --- Cloudiness (using S2 Cloud Probability dataset) ---
-    system_index = image.get("system:index")
-
-    cloud_img = (
-        ee.ImageCollection("COPERNICUS/S2_CLOUD_PROBABILITY")
-        .filter(ee.Filter.equals("system:index", system_index))
-        .first()
-    )
-
-    # Cloud probability band is "probability" (0–100)
-    cloud_prob = cloud_img.select("probability")
-    cloud_fraction = cloud_prob.reduceRegion(
-        reducer=ee.Reducer.mean(), geometry=region, scale=20, maxPixels=1e8
-    ).get("probability")
-
-    # --- Structure Presence (via NDWI min) ---
-    ndwi = b3.subtract(b8).divide(b3.add(b8))
-    ndwi_stats = ndwi.reduceRegion(
-        reducer=ee.Reducer.min(), geometry=region, scale=20, maxPixels=1e8
-    )
-    min_ndwi = ndwi_stats.get("B3")
-
-    structure_present = ee.Algorithms.If(
-        ee.Number(min_ndwi).lt(0),
-        1,  # structure
-        0,  # only ocean
-    )
-
-    # --- Flaring Detection (max difference B12 - B11) ---
-    # flare_diff = b12.subtract(b11).rename("flare_diff")
-    delta_sw = b12.subtract(b11)
-    tai = delta_sw.divide(b8a).rename("TAI")
-
-    # Get max value and pixel location
-    flare_reduce = tai.reduceRegion(
-        reducer=ee.Reducer.max(),
-        geometry=region,
-        scale=20,
-        maxPixels=1e8,
-        bestEffort=True,
-    )
-    max_flare_diff = flare_reduce.get("TAI")
-
-    # Get coordinates of max flare pixel
-    flare_mask = tai.eq(ee.Number(max_flare_diff))
-    flare_coords = (
-        flare_mask.reduceToVectors(
-            scale=20,
-            geometryType="centroid",
-            maxPixels=1e8,
-            bestEffort=True,
-        )
-        .filterBounds(region)
-        .geometry()
-        .coordinates()
-    )
-    flare_coords = ee.List(flare_coords)
-
-    flare_present = ee.Algorithms.If(
-        ee.Number(max_flare_diff).gt(0.15),  # threshold lowered
-        1,
-        0,
-    )
-
-    # Return as dictionary
-    result = ee.Dictionary(
-        {
-            "system_index": system_index,
-            "lat": lat,
-            "lon": lon,
-            "cloud_fraction": cloud_fraction,
-            "min_ndwi": min_ndwi,
-            "structure_present": structure_present,
-            "max_TAI": max_flare_diff,
-            "flare_present": flare_present,
-            "flare_latlon": flare_coords,
-        }
-    )
-
-    return result
-
-
 def get_image(sid):
     col = ee.ImageCollection("COPERNICUS/S2")
     return col.filter(ee.Filter.eq("system:index", sid)).first()
 
 
+def add_cloud_probability(s2_img):
+    # Get the granule ID from the system:index
+    granule_id = s2_img.get("system:index")
+
+    # Match with the cloud probability collection
+    s2_cloud = (
+        ee.ImageCollection("COPERNICUS/S2_CLOUD_PROBABILITY")
+        .filter(ee.Filter.eq("system:index", granule_id))
+        .first()
+    )
+
+    # Add the cloud probability band as "probability"
+    return s2_img.addBands(s2_cloud.rename("probability"))
+
+
 class ViewerState:
+    """
+    Stores information about the current view
+    """
+
     def __init__(self):
         self.sid = None  # will store currently displayed SID
 
@@ -234,6 +160,26 @@ def show_granule_viewer(
     starting_idx=0,
     layers=["B11", "B12", "RGB", "MBSP", "Flaring", "Detected Flare"],
 ):
+    """
+    Displays an interactive viewer for Sentinel-2 granules with flaring evaluation.
+
+    Parameters
+    ----------
+    sid_data : list of dict
+        Each dict must have a "SID" key and lat/lon.
+    b11_b12_max : list of float
+        Maximum values for B11 and B12 visualization (default: [3000, 2800]).
+    zoom : int
+        Initial zoom level for the map (default: 12).
+    mbsp_min_max : list of float
+        Minimum and maximum values for MBSP visualization (default: [-0.2, 0.2]).
+    extra_gdf : GeoDataFrame, optional
+        Additional GeoDataFrame to overlay on the map (default: None).
+    starting_idx : int
+        Index of the initial SID to display (default: 0).
+    layers : list of str
+        List of layers to display on the map (default: ["B11", "B12", "RGB", "MBSP", "Flaring", "Detected Flare"]).
+    """
     index = starting_idx
     state = ViewerState()
 
@@ -264,6 +210,7 @@ def show_granule_viewer(
         img = get_image(state.sid)
 
         region = ee.Geometry.Point([lon, lat]).buffer(200)
+        img = add_cloud_probability(img)
         img_flaring_data = evaluateScene(img, region).getInfo()
 
         vis_params_b11 = {"bands": ["B11"], "min": 0, "max": b11_b12_max[0]}
@@ -409,6 +356,19 @@ def show_granule_viewer(
 
 
 def gdf_to_sid_list(gdf):
+    """
+    Converts a GeoDataFrame with 'system_index' and 'geometry' columns to a list of dictionaries with 'SID', 'lat', and 'lon' keys.
+
+    Parameters
+    ----------
+    gdf : GeoDataFrame
+        A GeoDataFrame containing 'system_index' and 'geometry' columns.
+
+    Outputs
+    -------
+    sid_data : list of dict
+        A list of dictionaries, each containing 'SID', 'lat', and 'lon' keys corresponding to the system index and coordinates of each geometry in the GeoDataFrame.
+    """
     sid_data = []
     for i, row in gdf.iterrows():
         sid_data.append(
@@ -421,215 +381,32 @@ def gdf_to_sid_list(gdf):
     return sid_data
 
 
-def create_structure_flaring_timeline(
-    df, date_col="granule_date", filter_outlier=False, flare_override_absent=True
-):
-    df = df.copy()
-    df[date_col] = pd.to_datetime(df[date_col])
-    df = df.sort_values(by=date_col)
-
-    # Map to colors
-    def status_color_override(row):
-        if row["flaring"] == 1:
-            return "green"
-        elif row["structure_present"] == 0:
-            return "red"
-        else:
-            return "yellow"
-
-    def status_color(row):
-        if row["structure_present"] == 0:
-            return "red"
-        elif row["flaring"] == 1:
-            return "green"
-        else:
-            return "yellow"
-
-    if flare_override_absent:
-        df["color"] = df.apply(status_color_override, axis=1)
-    else:
-        df["color"] = df.apply(status_color, axis=1)
-
-    df["is_outlier"] = (df["color"] != df["color"].shift(1)) & (
-        df["color"] != df["color"].shift(-1)
-    )
-    if filter_outlier:
-        df = df[~df["is_outlier"]]
-    return df
-
-
-def plot_structure_flaring_timeline(
-    df, date_col="granule_date", magnitude_col=None, max_mag=1
-):
-    plt.figure(figsize=(12, 3))
-
-    if magnitude_col:
-        # Normalize lengths so they are visually balanced
-        lengths = df[magnitude_col].apply(
-            lambda h: max_mag if (h > max_mag or h < -max_mag) else h
-        )
-        # lengths = (vals / max_mag) * 0.8  # scale to 80% of axis height
-    else:
-        lengths = [0.8] * len(df)
-
-    for x, h, c in zip(df[date_col], lengths, df["color"]):
-        plt.vlines(x, 0, h, color=c, linewidth=2)
-
-    # Beautify
-    plt.title("Structure & Flaring Timeline", fontsize=14)
-    # plt.yticks([])
-    plt.xlabel("Date")
-    plt.grid(axis="x", linestyle="--", alpha=0.3)
-    plt.tight_layout()
-    plt.show()
-
-
-def extract_state_blocks(df, date_col="granule_date", drop_zero_day=True):
-    """
-    Collapse consecutive rows of the same 'color' into blocks.
-
-    Returns a dataframe with:
-      - start
-      - end
-      - color
-      - duration
-    """
-    color_to_state = {
-        "green": "flaring on",
-        "yellow": "flaring off",
-        "red": "structure absent",
-    }
-    df = df.copy().sort_values(date_col)
-    df = df.reset_index(drop=True)
-
-    # extras
-    # system_index = df['system_index'].iloc[0]
-    lat = df["geometry"].iloc[0].x
-    lon = df["geometry"].iloc[0].y
-    structure_id = df["structure_id"].iloc[0]
-
-    blocks = []
-    start = df.loc[0, date_col]
-    current_color = df.loc[0, "color"]
-    count = 1
-
-    for i in range(1, len(df)):
-        if df.loc[i, "color"] != current_color:
-            end = df.loc[i - 1, date_col]
-            blocks.append(
-                {
-                    "lat": lat,
-                    "lon": lon,
-                    "start": start,
-                    "end": end,
-                    "state": color_to_state[current_color],
-                    "color": current_color,
-                    "duration": end - start,
-                    "granule_count": count,
-                    # "system_index": system_index,
-                    "structure_id": structure_id,
-                }
-            )
-            # start new block
-            start = df.loc[i, date_col]
-            current_color = df.loc[i, "color"]
-            count = 1
-        else:
-            count += 1
-
-    # add final block
-    end = df.loc[len(df) - 1, date_col]
-    blocks.append(
-        {
-            "lat": lat,
-            "lon": lon,
-            "start": start,
-            "end": end,
-            "state": color_to_state[current_color],
-            "color": current_color,
-            "duration": end - start,
-            "granule_count": count,
-            # "system_index": system_index,
-            "structure_id": structure_id,
-        }
-    )
-    block_pd = pd.DataFrame(blocks)
-    if drop_zero_day:
-        block_pd = block_pd[block_pd["duration"] > timedelta(0)]
-
-    return block_pd
-
-
-def plot_state_blocks(blocks_df):
-    """
-    Plot timeline blocks with colors.
-    """
-    fig, ax = plt.subplots(figsize=(12, 2))
-
-    for _, row in blocks_df.iterrows():
-        ax.barh(
-            y=0,
-            width=(row["end"] - row["start"]).days
-            + (row["end"] - row["start"]).seconds / 86400,
-            left=row["start"],
-            color=row["color"],
-            edgecolor="black",
-            height=0.5,
-        )
-
-    ax.set_yticks([])
-    ax.set_xlabel("Date")
-    ax.set_title("State Blocks Timeline")
-    plt.grid(axis="x", linestyle="--", alpha=0.3)
-    plt.tight_layout()
-    plt.show()
-
-
-def assign_group_ids(gdf, buffer_size=0.1, geom_col="geometry", id_col="structure_id"):
-    gdf = gdf.copy()
-    gdf["buffer"] = gdf[geom_col].buffer(buffer_size)
-    gdf[id_col] = -1
-
-    group_id = 0
-    for idx, geom in gdf["buffer"].items():
-        if gdf.at[idx, id_col] == -1:  # not yet assigned
-            group_id += 1
-            intersects = gdf["buffer"].intersects(geom)
-            gdf.loc[intersects, id_col] = group_id
-
-    return gdf.drop(columns="buffer")
-
-
 # %%
-sid_data = [
-    {
-        "SID": "20170705T164319_20170705T165225_T15RXL",
-        "lon": -90.968,
-        "lat": 27.292,
-    },
-    {
-        "SID": "20230716T162841_20230716T164533_T15QWB",
-        "lon": -92.2361,
-        "lat": 19.56586,
-    },
-    {
-        "SID": "20240421T162841_20240421T164310_T15QWB",
-        "lon": -92.23655,
-        "lat": 19.56582,
-    },
-    {
-        "SID": "20220707T032531_20220707T033631_T48NTP",
-        "lon": 102.986983,
-        "lat": 7.592794,
-    },
-]
+# sid_data = [
+#     {
+#         "SID": "20170705T164319_20170705T165225_T15RXL",
+#         "lon": -90.968,
+#         "lat": 27.292,
+#     },
+#     {
+#         "SID": "20230716T162841_20230716T164533_T15QWB",
+#         "lon": -92.2361,
+#         "lat": 19.56586,
+#     },
+#     {
+#         "SID": "20240421T162841_20240421T164310_T15QWB",
+#         "lon": -92.23655,
+#         "lat": 19.56582,
+#     },
+#     {
+#         "SID": "20220707T032531_20220707T033631_T48NTP",
+#         "lon": 102.986983,
+#         "lat": 7.592794,
+#     },
+# ]
 
-b11 = 3000
 # state = show_granule_viewer(
 #     sid_data,
-#     b11_b12_max=[b11, b11],
-#     zoom=16,
-#     mbsp_min_max=[-0.1, 0.1],
 # )
 
 # %%
